Add rocket-compiler bin crate (#8)
* Add two annotation passes to the compiler plugin

* Add beginnings of rocket-compiler, wrapping the MLIR Transform dialect

* Update rocket-compiler report.rs

* Add op name to rocket-compiler report.rs

* Run cargo clippy --fix

+++ rocket-hal-driver/src/buffer.rs
@@ -99,7 +99,7 @@ unsafe extern "C" fn destroy(buffer: *mut iree_hal_buffer_t) {
         let _ = unsafe {
             nix::sys::mman::munmap(
                 std::ptr::NonNull::new_unchecked(buffer.host_ptr as *mut std::ffi::c_void),
-                buffer.base.allocation_size as usize,
+                buffer.base.allocation_size,
             )
         };
     }
